Move elgot and coelgot to refolds
